Atom: Hopefully fixed some bugs with threads, added clearer output to the signal handler
git-svn-id: https://svn.arune.se/svn/HomeAutomation@1124 e69ec2e4-db17-0410-bd4a-d6f28f89cd0d

+++ trunk/PcSoftware/Atom/src/main.cpp
@@ -35,6 +35,7 @@ using namespace std;
 #include "VM/virtualmachine.h"
 #include "CanNet/candebug.h"
 
+bool cleanUpRunning = false;
 int cleanUp();
 void handler(int status);
 
@@ -120,18 +121,65 @@ int cleanUp()
 {
 	Logger &log = Logger::getInstance();
 
-	log.add("\n");
-	log.addToSyslog("Thank you for using Atom. Goodbye!\n");
+	if (!cleanUpRunning)
+	{
+		cleanUpRunning = true;
+	
+		log.add("\n");
+		log.addToSyslog("Thank you for using Atom. Goodbye!\n");
 
-	CanDebug::deleteInstance();
-	VirtualMachine::deleteInstance();
-	CanNetManager::deleteInstance();
-	Logger::deleteInstance();
+		CanDebug::deleteInstance();
+		VirtualMachine::deleteInstance();
+		CanNetManager::deleteInstance();
+		Logger::deleteInstance();
+	}
+	else
+	{
+		log.addToSyslog("Something has gone wrong clean up is already running!\n");
+		return EXIT_FAILURE;
+	}
 
 	return EXIT_SUCCESS;
 }
 
 void handler(int status)
 {
+	string signalName = "Unknown";
+	
+	switch (status)
+	{
+	case SIGTERM:
+	signalName = "Terminate";
+	break;
+	
+	case SIGINT:
+	signalName = "Interupt";
+	break;
+	
+	case SIGQUIT:
+	signalName = "Quit";
+	break;
+	
+	case SIGABRT:
+	signalName = "Abort";
+	break;
+	
+	case SIGIO:
+	signalName = "I/O";
+	break;
+	
+	case SIGPIPE:
+	signalName = "Pipe";
+	break;
+	}
+
+	Logger &log = Logger::getInstance();
+	log.addToSyslog("Received signal " + signalName + "(" + itos(status) + ")\n");
+	
+	if (status == SIGPIPE)
+	{
+		return;
+	}
+	
 	exit(cleanUp());
 }
